+++ include/aspect/velocity_boundary_conditions/ascii_data.h
@@ -23,6 +23,8 @@
 #define __aspect__velocity_boundary_conditions_ascii_data_h
 
 #include <aspect/velocity_boundary_conditions/interface.h>
+#include <aspect/boundary_temperature/ascii_data.h>
+
 #include <aspect/simulator_access.h>
 #include <aspect/utilities.h>
 
@@ -40,7 +42,7 @@ namespace aspect
      * @ingroup VelocityBoundaryConditionsModels
      */
     template <int dim>
-    class AsciiData : public Utilities::AsciiDataBoundary<dim>, public Interface<dim>
+    class AsciiData : public BoundaryTemperature::internal::AsciiDataBoundary<dim>, public Interface<dim>
     {
       public:
         /**
