+++ source/velocity_boundary_conditions/ascii_data.cc
@@ -50,7 +50,7 @@ namespace aspect
       AssertThrow(*(boundary_ids.begin()) != numbers::invalid_boundary_id,
                   ExcMessage("Did not find the boundary indicator for the prescribed data plugin."));
 
-      Utilities::AsciiDataBoundary<dim>::initialize(boundary_ids,
+      BoundaryTemperature::internal::AsciiDataBoundary<dim>::initialize(boundary_ids,
                                                    dim);
     }
 
@@ -60,7 +60,7 @@ namespace aspect
     {
       Interface<dim>::update ();
 
-      Utilities::AsciiDataBoundary<dim>::update();
+      BoundaryTemperature::internal::AsciiDataBoundary<dim>::update();
     }
 
 
@@ -71,7 +71,7 @@ namespace aspect
     {
       Tensor<1,dim> velocity;
       for (unsigned int i = 0; i < dim;i++)
-        velocity[i] = Utilities::AsciiDataBoundary<dim>::get_data_component(*(boundary_ids.begin()),
+        velocity[i] = BoundaryTemperature::internal::AsciiDataBoundary<dim>::get_data_component(*(boundary_ids.begin()),
                                                                             position,
                                                                             i);
       return velocity;
@@ -84,7 +84,7 @@ namespace aspect
     {
       prm.enter_subsection("Boundary velocity model");
       {
-        Utilities::AsciiDataBoundary<dim>::declare_parameters(prm,
+        BoundaryTemperature::internal::AsciiDataBoundary<dim>::declare_parameters(prm,
                                                               "$ASPECT_SOURCE_DIR/data/velocity-boundary-conditions/ascii-data/test/",
                                                               "box_2d_%s.%d.txt");
       }
@@ -98,7 +98,7 @@ namespace aspect
     {
       prm.enter_subsection("Boundary velocity model");
       {
-        Utilities::AsciiDataBoundary<dim>::parse_parameters(prm);
+        BoundaryTemperature::internal::AsciiDataBoundary<dim>::parse_parameters(prm);
 
         if (this->convert_output_to_years() == true)
           {
